+++ src/experimaestro/scheduler/__init__.py
@@ -1,1 +1,17 @@
-from .base import *
+from .base import Scheduler, Listener
+from .workspace import Workspace
+from .experiment import experiment, FailedExperiment
+from .jobs import Job, JobState, JobFailureStatus, JobDependency, JobContext
+
+__all__ = [
+    "Scheduler",
+    "Listener",
+    "Workspace",
+    "experiment",
+    "FailedExperiment",
+    "Job",
+    "JobState",
+    "JobFailureStatus",
+    "JobDependency",
+    "JobContext",
+]
